filename changes; prepare to implement CLI
